Move assessment field mapping into helpers
A new helpers function has been created where the function of this
method better fits. This changes moves the placement of
`mapAssessmentConditionalFields` into helpers from the reference
data service.
